Make File and Math Handler
Added a make file, also added a as of yet non-working function that
should be able to take a string and return an int.  I have to eventually
make it work for the other variables

There are probably other changes not described here.  Started preparing
to put in the framework for math handler.  Fixed needing to include the
.cpp file in the .h file for ScriptHandler.